requirements.txt
EasyDict==1.7
opencv-python
Cython
scipy
pandas
pyyaml
json_tricks
scikit-image
tensorboardX
yacs
cffi
munkres
tqdm
torch
torchvision